# frontend/openapi/job-hunt.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionCreateRequest": {
        "properties": {
          "owner_token": {
            "maxLength": 512,
            "minLength": 32,
            "title": "Owner Token",
            "type": "string"
          }
        },
        "required": [
          "owner_token"
        ],
        "title": "SessionCreateRequest",
        "type": "object"
      },
      "SessionDeleteResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "SessionDeleteResponse",
        "type": "object"
      },
      "SessionResponse": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "local_date": {
            "format": "date",
            "title": "Local Date",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "owner_id",
          "display_name",
          "timezone",
          "local_date",
          "expires_at"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OwnerSessionCookie": {
        "description": "Opaque HttpOnly session issued by POST /api/session.",
        "in": "cookie",
        "name": "job_hunt_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Job Hunt Signal API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/session": {
      "delete": {
        "operationId": "delete_session_api_session_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OwnerSessionCookie": []
          }
        ],
        "summary": "Delete Session",
        "tags": [
          "owner-session"
        ]
      },
      "get": {
        "operationId": "get_session_api_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OwnerSessionCookie": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "owner-session"
        ]
      },
      "post": {
        "operationId": "create_session_api_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "owner-session"
        ]
      }
    },
...
  }
}